+++ src/GolfEngine/GameManagement/Entities/Sprite.cpp
