+++ src/GolfEngine/GameManagement/Tile.hpp
@@ -11,9 +11,9 @@
 
 #define TILE_SIZE 64
 
-#include "../Vector2.hpp"
-#include "../Renderables/Entity.hpp"
-#include "../Renderables/Renderable.hpp"
+#include "../Geometry/Vector2.hpp"
+#include "Entities/Entity.hpp"
+#include "../Rendering/Renderable.hpp"
 #include <SFML/Graphics.hpp>
 #include <vector>
