+++ src/GolfEngine/Rendering/RenderableVisitor.cpp
